@@ -18,16 +18,20 @@ publish_to: "none" # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  chewie: ^1.2.2
-  cupertino_icons: ^1.0.2
-  flutter_svg: ^0.22.0
+  image_picker: ^0.8.7+3  # for picking images
+  chewie: ^1.8.1
+  cupertino_icons: ^1.0.8
+  flutter_svg: ^2.0.10+1
   scroll_indicator: ^0.0.3
-  video_player: ^2.1.10
+  video_player: ^2.3.1
+  responsive_grid: ^2.4.4
+  mongol: ^8.0.0
+  intl: ^0.19.0
 
 dev_dependencies:
   flutter_test:
@@ -47,6 +51,7 @@ flutter:
     - assets/icons/
     - assets/images/
     - assets/videos/
+    - assets/fonts/
   # An image asset can refer to one or more resolution-specific "variants", see
   # https://flutter.dev/assets-and-images/#resolution-aware.
   # For details regarding adding assets from package dependencies, see
@@ -57,6 +62,9 @@ flutter:
   # list giving the asset and other descriptors for the font. For
   # example:
   fonts:
+    - family: TraditionalMongolian
+      fonts:
+        - asset: assets/fonts/mnglwritingotf.ttf
     - family: inter
       fonts:
         - asset: assets/fonts/inter/Inter-Regular.ttf
