@@ -1,12 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 
 class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
-  final Widget title;
-  final Widget leadingIcon;
-  final Function onPressedLeading;
-  final List<Widget> actions;
-  final Widget profilePicture;
-  final Function onPressedProfilePicture;
+  final Widget? title;
+  final Widget? leadingIcon;
+  final Function? onPressedLeading;
+  final List<Widget>? actions;
+  final Widget? profilePicture;
+  final Function? onPressedProfilePicture;
 
   CustomAppBar({this.title, this.leadingIcon, this.onPressedLeading, this.actions, this.profilePicture, this.onPressedProfilePicture});
 
@@ -16,18 +17,19 @@ class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
   @override
   Widget build(BuildContext context) {
     return AppBar(
-      brightness: Brightness.dark,
       backgroundColor: Colors.black,
       elevation: 0.0,
       centerTitle: true,
       title: title,
+      // ignore: unnecessary_null_comparison
       leading: (leadingIcon != null)
           ? IconButton(
-              icon: leadingIcon,
-              onPressed: onPressedLeading,
+              icon: leadingIcon!,
+              onPressed: (){onPressedLeading!();},
             )
           : null,
-      actions: (profilePicture != null) ? [_buildProfilePicture(profilePicture)] : actions,
+      actions: (profilePicture != null) ? [_buildProfilePicture(profilePicture!)] : actions,
+      systemOverlayStyle: SystemUiOverlayStyle.light,
     );
   }
 
@@ -35,7 +37,7 @@ class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
     return Padding(
       padding: const EdgeInsets.all(12),
       child: InkWell(
-        onTap: onPressedProfilePicture,
+        onTap: (){onPressedProfilePicture!();},
         borderRadius: BorderRadius.circular(60),
         child: ClipRRect(
           borderRadius: BorderRadius.circular(60),
