@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 
 class CustomTextField extends StatefulWidget {
-  final TextEditingController controller;
-  final Function onChanged;
+  final TextEditingController? controller;
+  final Function? onChanged;
   final String hintText;
   final String labelText;
   final EdgeInsetsGeometry padding;
@@ -12,8 +12,8 @@ class CustomTextField extends StatefulWidget {
   CustomTextField({
     this.controller,
     this.onChanged,
-    this.hintText,
-    this.labelText,
+    required this.hintText,
+    required this.labelText,
     this.obsecureText = false,
     this.margin = const EdgeInsets.only(bottom: 16),
     this.padding = const EdgeInsets.only(left: 20, bottom: 5, top: 8),
@@ -33,16 +33,16 @@ class _CustomTextFieldState extends State<CustomTextField> {
       decoration: BoxDecoration(borderRadius: BorderRadius.circular(10), color: Color(0xFF0E0E0E), border: Border.all(color: Color(0xFF494949), width: 1)),
       child: TextField(
         controller: widget.controller,
-        onChanged: widget.onChanged,
+        // onChanged: widget.onChanged,
         maxLines: 1,
         style: TextStyle(color: Colors.white, fontSize: 16, fontWeight: FontWeight.w400),
         obscureText: widget.obsecureText,
         decoration: InputDecoration(
           focusedBorder: InputBorder.none,
           hintText: widget.hintText,
-          hintStyle: TextStyle(color: Colors.white.withOpacity(0.5), height: 0, fontSize: 16),
+          hintStyle: TextStyle(color: Colors.white.withOpacity(0.5), height: 0, fontSize: 30),
           labelText: widget.labelText,
-          labelStyle: TextStyle(color: Colors.grey, height: 0, fontSize: 14, fontFamily: 'inter'),
+          labelStyle: TextStyle(color: Colors.grey, height: 0, fontSize: 26, fontFamily: 'inter'),
           floatingLabelBehavior: FloatingLabelBehavior.always,
         ),
       ),
