@@ -1,28 +1,120 @@
 import 'package:readky/model/core/news.dart';
 
 class NewsHelper {
-  static List<News> featuredNews =
-      featuredNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
-  static List<News> breakingNews =
-      breakingNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  // static List<News> featuredNews =
+      // featuredNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  static List<News> featuredNews = featuredNewsRawData.map((data) => News.fromJson(data)).toList();
+  static List<News> breakingNews = breakingNewsRawData.map((data) => News.fromJson(data)).toList();
+      // breakingNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
 
-  static List<News> recomendationNews =
-      recomendationNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  static List<News> recomendationNews = recomendationNewsRawData.map((data) => News.fromJson(data)).toList();
+      // recomendationNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
 
-  static List<News> allCategoriesNews =
-      allCategoriesNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  static List<News> allCategoriesNews = allCategoriesNewsRawData.map((data) => News.fromJson(data)).toList();
+      // allCategoriesNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
 
-  static List<News> searchNews =
-      searchNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  static List<News> searchNews = searchNewsRawData.map((data) => News.fromJson(data)).toList();
+      // searchNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
 
-  static List<News> bookmarkedNews =
-      bookmarkedNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
+  static List<News> bookmarkedNews = bookmarkedNewsRawData.map((data) => News.fromJson(data)).toList();
+      // bookmarkedNewsRawData.map((data) => News(title: data['title'], photo: data['photo'], date: data['date'], author: data['author'], description: data['description'])).toList();
 }
+
+    // 'photo': [
+    // 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    // 'https://i.ibb.co/FwC7yMJ/4c3ebafe4e3a.jpg'
+    // ],
 
 var featuredNewsRawData = [
   {
-    'title': '‘Space Jam: A New Legacy’ Takes on ‘Black Widow’ at the Box Office.',
-    'photo': 'assets/images/spacejam.jpg',
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': [
+      'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+      'https://i.ibb.co/FwC7yMJ/4c3ebafe4e3a.jpg',
+    ],
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'ᠬᠦᠰᠡᠯᠡᠩ'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'ᠬᠦᠰᠡᠯᠡᠩ'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'ᠵᠠᠷᠤᠳ ᠬᠦᠰᠡᠯᠡᠩ'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'ᠵᠠᠷᠤᠳ ᠬᠦᠰᠡᠯᠡᠩ'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'ABB anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
     'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
 
 It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
@@ -35,8 +127,120 @@ In a rare moment of streaming revenue transparency, Disney revealed last weekend
     'author': 'anon anon'
   },
   {
-    'title': 'How a Pivotal Scene in ‘The Green Knight’ Took Over a Year to Edit.',
-    'photo': 'assets/images/green knight.jpg',
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021 ',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
+    'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
+
+It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
+
+After “Black Widow” opened on July 9 to $80 million in North America, a record for COVID times, the comic book adaptation is projected to add $24 million to $28 million in its second weekend of release. Independent tracking services suggest that range is conservative and show the final weekend figure could reach $32 million through Sunday. A $30 million-plus haul in its second frame would be a decline on par with fellow Marvel standalone adventures such as 2018’s “Ant-Man and the Wasp” and 2017’s “Spider-Man: Homecoming,” which each dropped 60% from their initial weekends. It would also be a slight improvement on another pandemic-era release, Universal’s “Fast & Furious” sequel “F9,” which debuted to $60 million in June and plunged 67% in its sophomore outing.
+
+In a rare moment of streaming revenue transparency, Disney revealed last weekend that “Black Widow” collected $60 million globally on Disney Plus, where the film premiered (for a $30 surcharge) on the same day as its theatrical release. For now, Disney isn’t expected to update that figure beyond its opening weekend sales.
+    ''',
+    'date': 'Jul 14, 2021',
+    'author': 'anon anon'
+  },
+  {
+    'title': 'ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ ᠨᠢᠭᠡ ᠮᠤᠩᠭᠤᠯ ᠬᠤᠶᠠᠷ ᠮᠤᠩᠭᠤᠯ',
+    'photo': 'https://i.ibb.co/yqxdnNq/spacejam.jpg',
     'description': r'''Black Widow has faced down mad titans and genocidal robots, but can the veteran Avenger take down a band of scrappy Looney Toons?
 
 It’ll be a showdown between Disney’s and Marvel’s “Black Widow” and the Warner Bros. sequel “Space Jam: A New Legacy” at the domestic box office, with Scarlett Johansson’s superhero adventure poised to reclaim the top spot.
