@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:readky/route/slide_page_route.dart';
 import 'package:readky/view/screens/page_switch.dart';
 import 'package:readky/view/widgets/custom_text_field.dart';
+import 'package:mongol/mongol.dart';
 
 class LoginPage extends StatefulWidget {
   @override
@@ -15,9 +17,7 @@ class _LoginPageState extends State<LoginPage> {
     return Scaffold(
       backgroundColor: Colors.black,
       appBar: AppBar(
-        brightness: Brightness.dark,
         backgroundColor: Colors.black,
-        title: SvgPicture.asset('assets/icons/appname.svg'),
         centerTitle: true,
         leading: IconButton(
           icon: Icon(
@@ -27,7 +27,7 @@ class _LoginPageState extends State<LoginPage> {
           onPressed: () {
             Navigator.of(context).pop();
           },
-        ),
+        ), systemOverlayStyle: SystemUiOverlayStyle.light,
       ),
       body: GestureDetector(
         onTap: () => FocusManager.instance.primaryFocus?.unfocus(),
@@ -35,70 +35,83 @@ class _LoginPageState extends State<LoginPage> {
           shrinkWrap: true,
           physics: BouncingScrollPhysics(),
           children: [
-            // Section 1 - Welcome Title
-            Container(
-              margin: EdgeInsets.only(top: 30, bottom: 40),
-              padding: EdgeInsets.symmetric(horizontal: 20),
-              width: MediaQuery.of(context).size.width,
-              child: Text(
-                "Hii 👋 let's go back to reading. ",
-                style: TextStyle(color: Colors.white, fontWeight: FontWeight.w600, fontSize: 28, height: 150 / 100),
-              ),
-            ),
             // Section 2 - Form
             Container(
-              margin: EdgeInsets.only(bottom: 24),
+              margin: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width/3),
               width: MediaQuery.of(context).size.width,
-              padding: EdgeInsets.symmetric(horizontal: 20),
+              // padding: EdgeInsets.symmetric(horizontal: 20),
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.start,
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  CustomTextField(
-                    hintText: 'youremail@email.com',
-                    labelText: 'Email',
+                  SizedBox(
+                    height: 100,
+                    child: CustomTextField(
+                      hintText: 'youremail@email.com',
+                      labelText: 'Email', 
+                    ),
                   ),
-                  CustomTextField(
-                    labelText: 'Password',
-                    hintText: '********',
-                    obsecureText: true,
+                  SizedBox(
+                    height: 100,
+                    child: CustomTextField(
+                      labelText: 'Password',
+                      hintText: '********',
+                      obsecureText: true,
+                    ),
                   ),
                 ],
               ),
             ),
             // Section 3 - Register Button
             Container(
-              margin: EdgeInsets.symmetric(horizontal: 20),
-              width: MediaQuery.of(context).size.width,
-              height: 70,
-              child: ElevatedButton(
-                onPressed: () {
-                  Navigator.of(context).push(SlidePageRoute(child: PageSwitch()));
-                },
-                child: Text(
-                  'Log in',
-                  style: TextStyle(color: Colors.black, fontWeight: FontWeight.w600, fontSize: 16, fontFamily: 'inter'),
-                ),
-                style: ElevatedButton.styleFrom(
-                  primary: Colors.white,
-                  onPrimary: Colors.black,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(10),
-                  ),
-                ),
-              ),
-            ),
-            Container(
-              margin: EdgeInsets.only(top: 16),
-              width: MediaQuery.of(context).size.width,
-              alignment: Alignment.topCenter,
-              child: TextButton(
-                onPressed: () {},
-                child: Text('Forgot your password?'),
-                style: TextButton.styleFrom(
-                  primary: Colors.white.withOpacity(0.65),
-                  textStyle: TextStyle(fontWeight: FontWeight.w400),
-                ),
+              child: Row(
+                mainAxisAlignment: MainAxisAlignment.center,
+                children: [
+                    Container(
+                      margin: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width/15),
+                      child: MongolElevatedButton(
+                        onPressed: () {
+                          Navigator.of(context).push(SlidePageRoute(child: PageSwitch()));
+                        },
+                        child: MongolText(
+                          'ᠨᠡᠪᠲᠡᠷᠡᠬᠦ ', 
+                          style: TextStyle(fontFamily: 'TraditionalMongolian', color: Colors.purple),
+                          textScaleFactor: 4,
+                        ),
+                        style: ButtonStyle(
+                          backgroundColor: MaterialStateProperty.all<Color>(Colors.white),
+                          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                            RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(18.0),
+                            side: BorderSide(color: Colors.black)
+                            )
+                          ),
+                        )
+                      ),
+                    ),
+                    SizedBox(width: 30),
+                    Container(
+                      child: MongolElevatedButton(
+                        onPressed: () {
+                          
+                        },
+                        child: MongolText(
+                          'ᠨᠢᠭᠤᠣᠠ\nᠦᠭᠡ\nᠪᠠᠨ\nᠮᠠᠷᠲᠠᠪᠠ', 
+                          style: TextStyle(fontFamily: 'TraditionalMongolian', color: Colors.grey),
+                          textScaleFactor: 3,
+                        ),
+                        style: ButtonStyle(
+                          backgroundColor: MaterialStateProperty.all<Color>(Colors.white),
+                          shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                            RoundedRectangleBorder(
+                            borderRadius: BorderRadius.circular(18.0),
+                            side: BorderSide(color: Colors.black)
+                            )
+                          ),
+                        )
+                      ),
+                    ),
+                ],
               ),
             ),
           ],
