@@ -1,8 +1,10 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/services.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:readky/route/slide_page_route.dart';
 import 'package:readky/view/screens/page_switch.dart';
 import 'package:readky/view/widgets/custom_text_field.dart';
+import 'package:mongol/mongol.dart';
 
 class RegisterPage extends StatefulWidget {
   @override
@@ -15,9 +17,7 @@ class _RegisterPageState extends State<RegisterPage> {
     return Scaffold(
       backgroundColor: Colors.black,
       appBar: AppBar(
-        brightness: Brightness.dark,
         backgroundColor: Colors.black,
-        title: SvgPicture.asset('assets/icons/appname.svg'),
         centerTitle: true,
         leading: IconButton(
           icon: Icon(
@@ -27,7 +27,7 @@ class _RegisterPageState extends State<RegisterPage> {
           onPressed: () {
             Navigator.of(context).pop();
           },
-        ),
+        ), systemOverlayStyle: SystemUiOverlayStyle.light,
       ),
       body: GestureDetector(
         onTap: () => FocusManager.instance.primaryFocus?.unfocus(),
@@ -35,81 +35,65 @@ class _RegisterPageState extends State<RegisterPage> {
           shrinkWrap: true,
           physics: BouncingScrollPhysics(),
           children: [
-            // Section 1 - Welcome Title
-            Container(
-              margin: EdgeInsets.only(top: 30, bottom: 40),
-              padding: EdgeInsets.symmetric(horizontal: 20),
-              width: MediaQuery.of(context).size.width,
-              child: Text(
-                'Hello ! Register to get started 😁',
-                style: TextStyle(color: Colors.white, fontWeight: FontWeight.w600, fontSize: 28, height: 150 / 100, fontFamily: 'inter'),
-              ),
-            ),
             // Section 2 - Form
             Container(
-              margin: EdgeInsets.only(bottom: 24),
+              margin: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width/3),
               width: MediaQuery.of(context).size.width,
-              padding: EdgeInsets.symmetric(horizontal: 20),
+              // padding: EdgeInsets.symmetric(horizontal: 20),
               child: Column(
                 mainAxisAlignment: MainAxisAlignment.start,
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  CustomTextField(
-                    labelText: 'Full Name',
-                    hintText: 'Your Name',
+                  SizedBox(
+                    height: 100,
+                    child: CustomTextField(
+                      labelText: 'ᠳ᠋ᠤᠭᠠᠷ',
+                      hintText: 'ᠤᠲᠠᠰᠤᠨ ᠳ᠋ᠤᠭᠠᠷ',
+                    ),
                   ),
-                  CustomTextField(
-                    hintText: 'youremail@email.com',
-                    labelText: 'Email',
+                  SizedBox(
+                    height: 100,
+                    child: CustomTextField(
+                      labelText: 'Password',
+                      hintText: '********',
+                      obsecureText: true,
+                    ),
                   ),
-                  CustomTextField(
-                    labelText: 'Password',
-                    hintText: '********',
-                    obsecureText: true,
-                  ),
-                  CustomTextField(
-                    labelText: 'Re Password',
-                    obsecureText: true,
-                    hintText: '********',
+                  SizedBox(
+                    height: 100,
+                    child: CustomTextField(
+                      labelText: 'Re Password',
+                      obsecureText: true,
+                      hintText: '********',
+                    ),
                   ),
                 ],
               ),
             ),
             // Section 3 - Register Button
             Container(
-              margin: EdgeInsets.symmetric(horizontal: 20),
-              width: MediaQuery.of(context).size.width,
-              height: 70,
-              child: ElevatedButton(
+              margin: EdgeInsets.symmetric(horizontal: MediaQuery.of(context).size.width/2.1),
+              child: MongolElevatedButton(
                 onPressed: () {
                   Navigator.of(context).push(SlidePageRoute(child: PageSwitch()));
                 },
-                child: Text(
-                  'Register',
-                  style: TextStyle(color: Colors.black, fontWeight: FontWeight.w600, fontSize: 16),
+                child: MongolText(
+                  'ᠳᠠᠩᠰᠠᠯᠠᠬᠤ', 
+                  style: TextStyle(fontFamily: 'TraditionalMongolian', color: Color.fromARGB(255, 9, 5, 10)),
+                  textScaleFactor: 4,
                 ),
-                style: ElevatedButton.styleFrom(
-                  primary: Colors.white,
-                  onPrimary: Colors.black,
-                  shape: RoundedRectangleBorder(
-                    borderRadius: BorderRadius.circular(10),
+                style: ButtonStyle(
+                  backgroundColor: MaterialStateProperty.all<Color>(Color.fromARGB(255, 189, 171, 12)),
+                  // backgroundColor: Colors.white, // This is what you need!
+                  shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                    RoundedRectangleBorder(
+                    borderRadius: BorderRadius.circular(18.0),
+                    side: BorderSide(color: Colors.black)
+                    )
                   ),
-                ),
+                )
               ),
             ),
-            Container(
-              margin: EdgeInsets.only(top: 16),
-              width: MediaQuery.of(context).size.width,
-              alignment: Alignment.topCenter,
-              child: TextButton(
-                onPressed: () {},
-                child: Text('Already have an account? Login'),
-                style: TextButton.styleFrom(
-                  primary: Colors.white.withOpacity(0.65),
-                  textStyle: TextStyle(fontWeight: FontWeight.w400),
-                ),
-              ),
-            )
           ],
         ),
       ),
