@@ -1,9 +1,12 @@
+import 'dart:html';
+
 import 'package:flutter/material.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:readky/route/slide_page_route.dart';
 import 'package:readky/view/screens/auth/login_page.dart';
 import 'package:readky/view/screens/auth/register_page.dart';
 import 'package:readky/view/widgets/switchable_button.dart';
+import 'package:mongol/mongol.dart';
 
 class WelcomePage extends StatefulWidget {
   @override
@@ -15,19 +18,6 @@ class _WelcomePageState extends State<WelcomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       backgroundColor: Colors.black,
-      bottomNavigationBar: Container(
-        padding: EdgeInsets.symmetric(vertical: 30),
-        width: MediaQuery.of(context).size.width,
-        child: SwitchableButton(
-          backgroundColor: Colors.white.withOpacity(0.1),
-          onTapLeft: () {
-            Navigator.of(context).push(SlidePageRoute(child: RegisterPage()));
-          },
-          onTapRight: () {
-            Navigator.of(context).push(SlidePageRoute(child: LoginPage()));
-          },
-        ),
-      ),
       body: Container(
         width: MediaQuery.of(context).size.width,
         height: MediaQuery.of(context).size.height,
@@ -42,30 +32,60 @@ class _WelcomePageState extends State<WelcomePage> {
                 color: Colors.black,
                 alignment: Alignment.topCenter,
                 child: Image.asset(
-                  'assets/images/welcomepagebg.jpg',
-                  fit: BoxFit.fitWidth,
-                  width: MediaQuery.of(context).size.width,
+                  'images/webhome.jpg',
                 ),
               ),
               Positioned(
-                top: MediaQuery.of(context).size.height * (48 / 100),
+                top: MediaQuery.of(context).size.height / 2,
+                left: MediaQuery.of(context).size.width / 2 - 85,
                 child: Container(
                   width: MediaQuery.of(context).size.width,
-                  height: 100,
-                  child: Column(
+                  height: 250,
+                  child: Row(
                     children: [
-                      SvgPicture.asset(
-                        'assets/icons/appname.svg',
-                        width: MediaQuery.of(context).size.width * (3 / 10),
-                      ),
-                      Container(
-                        margin: EdgeInsets.only(top: 18),
-                        child: Text(
-                          'Free Flutter News App \nStarter Template',
-                          style: TextStyle(color: Colors.white, height: 150 / 100, fontWeight: FontWeight.w500, fontSize: 16),
-                          textAlign: TextAlign.center,
+                     
+                        MongolElevatedButton(
+                          onPressed: () {
+                            Navigator.push(context, MaterialPageRoute(builder: (BuildContext context) => RegisterPage()));
+                          },
+                          child: MongolText(
+                            'ᠳᠠᠩᠰᠠᠯᠠᠬᠤ', 
+                            style: TextStyle(fontFamily: 'TraditionalMongolian', color: Color.fromARGB(255, 9, 5, 10)),
+                            textScaleFactor: 4,
+                          ),
+                          style: ButtonStyle(
+                            backgroundColor: MaterialStateProperty.all<Color>(Color.fromARGB(255, 189, 171, 12)),
+                            // backgroundColor: Colors.white, // This is what you need!
+                            shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                              RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(18.0),
+                              side: BorderSide(color: Colors.black)
+                              )
+                            ),
+                          )
                         ),
-                      )
+                      
+                      SizedBox(width: 30),
+                      MongolElevatedButton(
+                          onPressed: () {
+                            Navigator.push(context, MaterialPageRoute(builder: (BuildContext context) => LoginPage()));
+                          },
+                          child: MongolText(
+                            'ᠨᠡᠪᠲᠡᠷᠡᠬᠦ ', 
+                            style: TextStyle(fontFamily: 'TraditionalMongolian', color: Colors.purple),
+                            textScaleFactor: 4,
+                          ),
+                          style: ButtonStyle(
+                            backgroundColor: MaterialStateProperty.all<Color>(Colors.white),
+                            // backgroundColor: Colors.white, // This is what you need!
+                            shape: MaterialStateProperty.all<RoundedRectangleBorder>(
+                              RoundedRectangleBorder(
+                              borderRadius: BorderRadius.circular(18.0),
+                              side: BorderSide(color: Colors.black)
+                              )
+                            ),
+                          )
+                      ),
                     ],
                   ),
                 ),
